server/requirements.txt
Django==3.2.8
djangorestframework==3.12.4
Jinja2==3.0.2
mysqlclient==2.1.0
django-allauth==0.46.0
django-filter==21.1
jsonschema=3.2.0
